изменена структура проекта в соответсвии с  maven-archetype-webapp
