--- hw_cookies/.settings/org.eclipse.wst.common.project.facet.core.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<faceted-project>
-  <runtime name="Apache Tomcat v7.0"/>
-  <fixed facet="jst.web"/>
-  <fixed facet="java"/>
-  <fixed facet="wst.jsdt.web"/>
-  <installed facet="jst.web" version="3.0"/>
-  <installed facet="wst.jsdt.web" version="1.0"/>
-  <installed facet="java" version="1.8"/>
-</faceted-project>
