--- hw_cookies/.settings/org.eclipse.wst.jsdt.ui.superType.name
@@ -1,1 +1,0 @@
-Window
