+++ hw_cookies/pom.xml
@@ -25,22 +25,6 @@
 			<scope>test</scope>
 		</dependency>
 
-
-
-		<dependency>
-			<groupId>org.apache.tomcat</groupId>
-			<artifactId>tomcat-jsp-api</artifactId>
-			<version>7.0.34</version>
-		</dependency>
-
-		<!-- https://mvnrepository.com/artifact/javax.servlet/jsp-api -->
-		<dependency>
-			<groupId>javax.servlet</groupId>
-			<artifactId>jsp-api</artifactId>
-			<version>2.0</version>
-			<scope>provided</scope>
-		</dependency>
-
 		<!-- https://mvnrepository.com/artifact/javax.servlet/javax.servlet-api -->
 		<dependency>
 			<groupId>javax.servlet</groupId>
